# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recipes API",
        "description": "API documentation for the Recipes application",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/favorites": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Get user's favorite recipes",
                "operationId": "bc433e786ab5a0bec81c53c578e0b6aa",
                "responses": {
                    "200": {
                        "description": "List of favorite recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Spaghetti Carbonara"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Classic Italian pasta dish"
                                            },
                                            "ingredients": {
                                                "type": "string",
                                                "example": "Pasta, eggs, cheese, pancetta"
                                            },
                                            "instructions": {
                                                "type": "string",
                                                "example": "1. Cook pasta..."
                                            },
                                            "cooking_time": {
                                                "type": "integer",
                                                "example": 30
                                            },
                                            "servings": {
                                                "type": "integer",
                                                "example": 4
                                            },
                                            "difficulty": {
                                                "type": "string",
                                                "example": "medium"
                                            },
                                            "category": {
                                                "type": "string",
                                                "example": "Italian"
                                            },
                                            "image_url": {
                                                "type": "string",
                                                "example": "http://example.com/image.jpg"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "is_favorite": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "average_rating": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 4.5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/recipes/{id}/favorite": {
            "get": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Check if recipe is favorited",
                "operationId": "6d00ef8f6b3e896dbbdba1d5db1b993d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recipe ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favorite status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "is_favorite": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Add a recipe to favorites",
                "operationId": "de8ec7d1b58dabcfcf94918fe6651a85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recipe ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe added to favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe added to favorites"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Remove a recipe from favorites",
                "operationId": "e505807e9c661e8165417d7eac6f1bf7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recipe ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe removed from favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recipe removed from favorites"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recipe not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}